naieve simulation for quadtree

+++ immagine/include/Immagine/Imagedef.h
@@ -47,8 +47,8 @@ typedef struct TAG_COLOR
 
 typedef struct TAG_REGION
 {
-	uint16_t x;
-	uint16_t y;
-	uint16_t width;
-	uint16_t height;
+	size_t x;
+	size_t y;
+	size_t width;
+	size_t height;
 }Region;
